refactor to ConfigHelpers
